+++ src/Memory.h
@@ -18,8 +18,8 @@ public:
 
 	//Get RDRAM
 	unsigned char*    getRDRAM(int address=0)                { return &m_RDRAM[address];          }
-	unsigned long* getRDRAMint32()                        { return (unsigned long*)m_RDRAM; }
-	unsigned long     getRDRAMSize()                         { return m_RDRAMSize;                }
+	unsigned int* getRDRAMint32()                        { return (unsigned int*)m_RDRAM; }
+	unsigned int     getRDRAMSize()                         { return m_RDRAMSize;                }
 
 	//Get DMEM
 	unsigned char* getDMEM(int address=0)                    { return &m_DMEM[address];           }
@@ -28,12 +28,12 @@ public:
 	static unsigned long long* getTextureMemory(int address=0) { return &m_TMEM[address];           }
 
 	//Get Segment adress
-	unsigned long getRDRAMAddress(unsigned int segmentAddress) 
+	unsigned int getRDRAMAddress(unsigned int segmentAddress) 
 	{ 		
 		return (m_segments[(segmentAddress >> 24) & 0x0F] + (segmentAddress & 0x00FFFFFF)) & 0x00FFFFFF; 
 	}
 
-	void setSegment(unsigned long address, unsigned long value)
+	void setSegment(unsigned int address, unsigned int value)
 	{
 		if ( address >= 16 ) {
 			//ERROR
@@ -53,8 +53,8 @@ private:
 	unsigned char*          m_RDRAM;          //!< Rambus Dynamic Random Access Memory
 	unsigned char*          m_DMEM;           //!< RSP Data Memory
 	static unsigned long long m_TMEM[512];      //!< Texture Memory		
-	unsigned long        m_segments[16];   //!< Temporary memory for storing segment values
-	unsigned long           m_RDRAMSize;      //!< Size of RDRAM
+	unsigned int        m_segments[16];   //!< Temporary memory for storing segment values
+	unsigned int           m_RDRAMSize;      //!< Size of RDRAM
 
 };
